% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_bias.R
\name{bias}
\alias{bias}
\alias{bias.data.frame}
\alias{bias_vec}
\title{Bias}
\usage{
bias(data, ...)

\method{bias}{data.frame}(data, truth, estimate, na_rm = TRUE, ...)

bias_vec(truth, estimate, na_rm = TRUE, ...)
}
\arguments{
\item{data}{A \code{data.frame} containing the columns specified by the \code{truth}
and \code{estimate} arguments.}

\item{...}{Not currently used.}

\item{truth}{The column identifier for the true results
(that is \code{numeric}). This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column
names). For \verb{_vec()} functions, a \code{numeric} vector.}

\item{estimate}{The column identifier for the predicted
results (that is also \code{numeric}). As with \code{truth} this can be
specified different ways but the primary method is to use an
unquoted variable name. For \verb{_vec()} functions, a \code{numeric} vector.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA}
values should be stripped before the computation proceeds.}
}
\description{
A forecast bias occurs when there are consistent differences between actual
outcomes and previously generated forecasts of those quantities; that is:
forecasts may have a general tendency to be too high or too low. A normal
property of a good forecast is that it is not biased.
}
\details{
For reference: https://en.wikipedia.org/wiki/Forecast_bias
}
\author{
Florian Stracke
}
