% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celverr-package.R
\docType{package}
\name{celverr-package}
\alias{celverr}
\alias{celverr-package}
\title{celverr: Usefull collections of R functions from celver}
\description{
Some useful functions out of our projects. Focus is mainly on forecasting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/flostracke/celverr}
  \item Report bugs at \url{https://github.com/flostracke/celverr/issues}
}

}
\author{
\strong{Maintainer}: Florian Stracke \email{f.stracke@celver.com}

}
\keyword{internal}
