% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{check_colnames_present}
\alias{check_colnames_present}
\title{Helper function which can check if the needed columns are inside a Dataframe.}
\usage{
check_colnames_present(df, expected_col_names)
}
\arguments{
\item{df}{A Dataframe. which should be checked.}

\item{expected_col_names}{A list of the column names which should be in the
Dataframe.}
}
\value{
Returns a Error Message if the the columns is not in the Dataframe.
}
\description{
Helper function which can check if the needed columns are inside a Dataframe.
}
