% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{check_colnames_resample_results}
\alias{check_colnames_resample_results}
\title{Check a data frame containing forecasts against a resample set.}
\usage{
check_colnames_resample_results(df)
}
\arguments{
\item{df}{A Dataframe. which should be checked.}
}
\value{
Returns a Error Message if the the columns is not in the Dataframe.
}
\description{
Check a data frame containing forecasts against a resample set.
}
