% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_best_method.R
\name{choose_best_method}
\alias{choose_best_method}
\title{Choose per group the best performing forecast method.}
\usage{
choose_best_method(
  forecasts_resamples,
  forecasts_oos,
  metric,
  group_col = series_id
)
}
\arguments{
\item{forecasts_resamples}{The forecasts against the resample set.}

\item{forecasts_oos}{The forecasts against the out of sample set.}

\item{metric}{The metric used for deciding which is the best method. Has to
be a \code{yardstick} metric.}

\item{group_col}{The column on which level the decision should be made.
Default = series_id}
}
\description{
Assuming we created forecasts with multiple different algorithms we want
to choose the best performing algorithm based on a metric, calculated from
the resample set. The forecast then chooses the out of sample forecasts. Both the choosen
forecasts against the resample set and the out of sample set are returned
in a list.
}
