% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_misc.R
\name{combine_splits_and_preds}
\alias{combine_splits_and_preds}
\title{Combine the validation forecasts with the split object.}
\usage{
combine_splits_and_preds(split_obj, fit_resamples_res)
}
\arguments{
\item{split_obj}{A split object created by
\code{modeltime.resample::time_series_cv}}

\item{fit_resamples_res}{The fitted resamples produced by
````modeltime::modeltime_fit_resamples``}
}
\value{
The combined forecasts with the split object.
}
\description{
Combine the validation forecasts with the split object.
}
