% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_misc.R
\name{export_cv_to_tibble}
\alias{export_cv_to_tibble}
\title{Exports a CV plan as tibble. Useful if a package is not integrated
into the modeltime framework or for usage in Python.}
\usage{
export_cv_to_tibble(cv_plan)
}
\arguments{
\item{cv_plan}{A cross validation plan generated by \code{modeltime.resample::time_series_cv}}
}
\value{
A tibble containing the cross validation plan. The id columns stores the Slice information. The dataset contains the information if it is the train part (dataset = "train") or if it is the part which should be forecasted (dataset = "validation").
}
\description{
Exports a CV plan as tibble. Useful if a package is not integrated
into the modeltime framework or for usage in Python.
}
