% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_misc.R
\name{fix_model_names}
\alias{fix_model_names}
\title{Cleans some names of some forecasting models.}
\usage{
fix_model_names(x)
}
\arguments{
\item{x}{The dataframe with the produced forecasts.}
}
\value{
The dataframe with the cleaned model names.
}
\description{
The ETS models and the TBATS models don't have consistet naming. The f
unctions ensures that all ETS models are called ETS and all BATS models are
called TBATS.
}
