% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_misc.R
\name{guarante_output}
\alias{guarante_output}
\title{Ensures that forecasts are in the correct data format. Only the needed
columns are selected.}
\usage{
guarante_output(x, type = "validation")
}
\arguments{
\item{x}{The dataframe with the produced forecasts.}

\item{type}{Can be "validation" or "forecast". There are minor differences
in the expected columns between the 2 types.}
}
\value{
The dataframe with the needed columns
}
\description{
We have some assumptions about column names produced by the forecasts for
downstream useage. This function checks the data format.
}
