% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_best_method.R
\name{plot_best_choosen_method}
\alias{plot_best_choosen_method}
\title{Plots the choosen models based on the validation ouput of \code{choose_best_method}}
\usage{
plot_best_choosen_method(best_method_val_results, subtitle = "")
}
\arguments{
\item{best_method_val_results}{The output validation dataframe of \code{choose_best_method}}

\item{subtitle}{specifiy the subtitle for the plot}
}
\description{
Plots the choosen models based on the validation ouput of \code{choose_best_method}
}
