% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsle.R
\name{rmsle}
\alias{rmsle}
\title{Root mean squared log error}
\usage{
rmsle(data, ...)
}
\arguments{
\item{data}{The data frame containing the predictions and the actual values}

\item{...}{currently not used}
}
\value{
the calculated metric
}
\description{
Calculate the root mean squared log error. In the case of RMSE, the presence
of outliers can explode the error term to a very high value. But, in the
case of RMLSE the outliers are drastically scaled down therefore nullifying
their effect. RMSLE incurs a larger penalty for the underestimation of the
Actual variable than the Overestimation. This is especially useful for
business cases where the underestimation of the target variable is not
acceptable but overestimation can be tolerated.
}
\details{
For reference check this explanation: https://medium.com/analytics-vidhya/root-mean-square-log-error-rmse-vs-rmlse-935c6cc1802a
}
