% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsle.R
\name{rmsle_vec}
\alias{rmsle_vec}
\title{Root Mean Squared Log Error}
\usage{
rmsle_vec(truth, estimate, na_rm = TRUE, ...)
}
\arguments{
\item{truth}{The column identifier for the true results
(that is \code{numeric}). This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column
names). For \verb{_vec()} functions, a \code{numeric} vector.}

\item{estimate}{The column identifier for the predicted
results (that is also \code{numeric}). As with \code{truth} this can be
specified different ways but the primary method is to use an
unquoted variable name. For \verb{_vec()} functions, a \code{numeric} vector.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA}
values should be stripped before the computation proceeds.}

\item{...}{Not currently used.}

\item{data}{A \code{data.frame} containing the columns specified by the \code{truth}
and \code{estimate} arguments.}
}
\description{
Same like \code{rmsle} but for vectors.
}
\author{
Florian Stracke
}
