% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_misc.R
\name{stabilize_forecasts}
\alias{stabilize_forecasts}
\title{Averages non determistic forecast models}
\usage{
stabilize_forecasts(model_table, n_iterations = 10)
}
\arguments{
\item{model_table}{A modeltime table with the trained models.}

\item{n_iterations}{Number of how many forecasts are averaged.}
}
\value{
The summarised forecast as a tibble.
}
\description{
If the created forecasts of a forecast function are not deterministic this function creates a mean ensemble of \code{n_iterations}.
}
