% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_wape.R
\name{wape}
\alias{wape}
\alias{wape.data.frame}
\alias{wape_vec}
\title{wape}
\usage{
wape(data, ...)

\method{wape}{data.frame}(data, truth, estimate, na_rm = TRUE, ...)

wape_vec(truth, estimate, na_rm = TRUE, ...)
}
\arguments{
\item{data}{A \code{data.frame} containing the columns specified by the \code{truth}
and \code{estimate} arguments.}

\item{...}{Not currently used.}

\item{truth}{The column identifier for the true results
(that is \code{numeric}). This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column
names). For \verb{_vec()} functions, a \code{numeric} vector.}

\item{estimate}{The column identifier for the predicted
results (that is also \code{numeric}). As with \code{truth} this can be
specified different ways but the primary method is to use an
unquoted variable name. For \verb{_vec()} functions, a \code{numeric} vector.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA}
values should be stripped before the computation proceeds.}
}
\description{
The Weighted Absolute Percentage Error (WAPE) measures the overall deviation
of forecasted values from observed values. WAPE is calculated by taking the
sum of observed values and the sum of predicted values, and calculating the
error between those two values. A lower value indicates a more accurate model.
When the sum of observed values for all time points and all items is
approximately zero in a given backtest window, the weighted absolute
percentage error expression is undefined. In these cases, Forecast outputs
the unweighted absolute error sum, which is the numerator in the WAPE
expression. WAPE is more robust to outliers than
Root Mean Square Error (RMSE) because it uses the absolute error instead of
the squared error.
}
\details{
For reference: https://www.baeldung.com/cs/mape-vs-wape-vs-wmape or
https://docs.aws.amazon.com/forecast/latest/dg/metrics.html#metrics-WAPE
}
\author{
Florian Stracke
}
