% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{fcst_local_models_groups}
\alias{fcst_local_models_groups}
\title{Title}
\usage{
fcst_local_models_groups(
  modellist,
  cv_plan,
  forecast_horizon,
  train_data,
  use_parallel = TRUE,
  inner_loop_var
)
}
\arguments{
\item{inner_loop_var}{}
}
\description{
Title
}
